fix: resolve import cycle by moving TenantAwareAuthService to services package
- Move tenant_aware.go and tests from auth to services package
- Rename TenantAwareService to TenantAwareAuthService
- Add logger parameter to constructor
- Update Makefile lint command to use golangci-lint directly
- Fix error checking in repository tests
- Remove trailing newlines per gofmt

This resolves the import cycle between auth and services packages
that was preventing compilation.
