+++ pkg/models/tenant_config_test.go
@@ -378,4 +378,3 @@ func TestTenantConfig_CompleteScenario(t *testing.T) {
 
 	assert.Contains(t, tc.AllowedOrigins, "https://app.example.com")
 }
-
